requirements.txt
opencv-python>=4.0
numpy>=1.20
mss
pillow
pywebview
loguru